Restrict discoverHeartManDevices() to actual HeartMan's
This change adds a check for the name of the bluetooth device. If it
contains the string "heartman", it will be added to the list, returned
by the discoverHeartManDevices() method. This check assures that only
HeartMan devices are returned and all other Bluetooth devices in the
environment are ignored.

This closes #8